move SimpleDataBindingAdapter to gto-support-androidx-recyclerview
