+++ gto-support-recyclerview/src/main/kotlin/org/ccci/gto/android/common/recyclerview/adapter/SimpleDataBindingAdapter.kt
@@ -2,16 +2,8 @@ package org.ccci.gto.android.common.recyclerview.adapter
 
 import androidx.databinding.ViewDataBinding
 import androidx.lifecycle.LifecycleOwner
-import org.ccci.gto.android.common.androidx.recyclerview.adapter.AbstractDataBindingAdapter
-import org.ccci.gto.android.common.androidx.recyclerview.adapter.DataBindingViewHolder
+import org.ccci.gto.android.common.androidx.recyclerview.adapter.SimpleDataBindingAdapter
 
+@Deprecated("Since v3.11.2, use SimpleDataBindingAdapter from gto-support-androidx-recyclerview instead.")
 abstract class SimpleDataBindingAdapter<B : ViewDataBinding>(lifecycleOwner: LifecycleOwner? = null) :
-    AbstractDataBindingAdapter<B, DataBindingViewHolder<B>>(lifecycleOwner) {
-    final override fun onCreateViewHolder(binding: B, viewType: Int) = DataBindingViewHolder(binding)
-    final override fun onBindViewHolder(holder: DataBindingViewHolder<B>, binding: B, position: Int) =
-        onBindViewDataBinding(binding, position)
-    final override fun onViewRecycled(holder: DataBindingViewHolder<B>, binding: B) = onViewDataBindingRecycled(binding)
-
-    protected abstract fun onBindViewDataBinding(binding: B, position: Int)
-    protected open fun onViewDataBindingRecycled(binding: B) = Unit
-}
+    SimpleDataBindingAdapter<B>(lifecycleOwner)
